Finished tests for environment install procedure
