+++ tests/README.md
@@ -1,0 +1,17 @@
+# Unit Tests
+
+## Tests Defined
+
+***
+
+* test_db_setup.py (Tests the db environment installation procedure)
+* test_db_teardown.py (Tests the db environment teardown procedure)
+
+## Usage
+
+***
+
+Each unittest can be run by navigating to the root directory of this project
+and typing the following:
+
+> python3 -m unittest <test_name>
