+++ todo/tests.md
@@ -1,2 +1,2 @@
-Tests that still need to be written
+# Tests that still need to be written/finished
 * test_account_unique_email (Test for unique email insertion)
